chore: decouple formatting-related methods
